VarianceComponentsHDFE

This package runs Kline, Saggio and Sølvsten (KSS) bias correction of variance components in two-way fixed effects models. This is achieved by running an executable (app) where the user only needs to input the original data (in .csv format) that contains a first identier (e.g. worker id), a second identier (e.g firm id) and outcome (e.g. log wage). The link to the repository is this.

About the executable/app

The algorithm prints the plug-in and the bias-corrected variance components estimators for the first identifier effects (e.g. variance of worker effects), the second identifier effects (e.g. variance of firm effects), and the covariance of both these effects (e.g. covariance of worker-firm effects). The user may choose to compute only a subset of these three components. Additionally, the executable will create a CSV file that stores outcome, ids, and the fixed effects for every observation belonging to the leave-out connected set as defined in KSS. The algorithm also saves the statistical leverages, Pii, as well the weighting terms defined as Bii in KSS.

Installation (Windows)

Open up a powershell terminal (Windows + R, then type "powershell"), and press Enter.

Change the current directory by typing cd "your_desired_path".

Hint : To copy-paste into the terminal use the standard Ctrl+C and paste into the powershell by using right click.

In the powershell, install the latest version:

wget https://github.com/HighDimensionalEconLab/VarianceComponentsHDFE.jl/releases/download/v0.1.5/vchdfe--windows-latest.tar.gz -O vchdfe-windows-latest.tar.gz
 
tar -xvf vchdfe-windows-latest.tar.gz

Note that to be able to use wget on Windows Internet Explorer should have been launched at least once.

(OPTIONAL): Add the current directory to PATH:

setx PATH "$env:path;$pwd\vchdfe\bin" -m

Note: To permanently change the path, you need to run powershell as adminstrator.

Test using the sample test file:

vchdfe vchdfe\bin\test.csv

Executable guide

Use

vchdfe --help

to see a complete list of available arguments:

positional arguments:
  path                  path to CSV file containing data

optional arguments:
  --first_id FIRST_ID   column index in CSV file for the first ID
                        (e.g. Person).  Use the most granular type.
                        (type: Int64, default: 1)
  --second_id SECOND_ID
                        column index in CSV file for the second ID
                        (e.g. Firm).  Use the less granular type.
                        (type: Int64, default: 2)
  --observation_id OBSERVATION_ID
                        column index in CSV file for observation (e.g.
                        Wage). (type: Int64, default: 4)
  --first_id_effects    Computing and showing first_id effects
  --cov_effects         Computing and showing covariace effects
  --algorithm ALGORITHM
                        type of algorithm: Exact or JLA. It defaults
                        to be Exact if the number of observations is
                        less than 5000, and JLA otherwise. (default:
                        "Default")
  --simulations SIMULATIONS
                        number of simulations in the JLA algorithm. If
                        0, defaults to 100 * log(#total fixed effect)
                        (type: Int64, default: 0)
  --header              CSV file contains header
  --first_id_display FIRST_ID_DISPLAY
                        The display text associated with first_id
                        (e.g. Person). (default: "Person")
  --second_id_display SECOND_ID_DISPLAY
                        The display text associated with second_id
                        (e.g. Firm) (default: "Firm")
  --observation_id_display OBSERVATION_ID_DISPLAY
                        The display text associated with observable_id
                        (e.g. Wage) (default: "Wage")
  --detailed_output_path DETAILED_OUTPUT_PATH
                        path to the CSV for the detailed output for
                        each observable (default:
                        "C:\\Users\\owner\\Desktop\\vchdfe\\variance_components.csv")
  --results_path RESULTS_PATH
                        path to the results of the output (default:
                        "C:\\Users\\owner\\Desktop\\vchdfe\\results.txt")
  --write_detailed_CSV  write the detailed output to a CSV
  --write_results       write the results to a file
  --print_level PRINT_LEVEL
                        Level of verbosity of output. (type: Int64,
                        default: 1)
  -h, --help            show this help message and exit

A detailed list of the output is provided below. The program provides two outputs: an output file contains the main results and a detailed output file.

Results output

By using the argument --write_results followed by --results_path MYFILE, you can specify that you want to write the results in MYFILE.

The output file is a .txt file including

Detailed output

You can save all the output details in a CSV file. You need to use --write_detailed_CSV followed by --detailed_output_path MYFILE, where MYFILE is the path to the detailed output file.

The detailed output file includes:

observation : observation identifier in the original dataset that belong to the Leave-out connected set. For instance, if the number 3 appears in this vector, it means that the third observation in the original dataset belongs to the leave-out connected set.

first_id: the first identifier corresponding to each observation in obs (e.g. worked ids in the leave-out connected set).

second_id: the second identifier corresponding to each observation in obs (e.g. firm ids in the leave-out connected set).

beta: the vector of coefficients

D_alpha: the fixed effect for the first identifier corresponding to each observation. We can obtain the plug-in estimator of the variance component of this effect by computing the variance of this vector.

F_psi: the fixed effect for the second identifier corresponding to each observation. We can obtain the plug-in estimator of the variance component of this effect by computing the variance of this vector.

Pii: statistical leverage corresponding to each observation in obs.

Bii_first: The Bii to correct the variance of first_id effects corresponding to each observation in obs.

Bii_second: The Bii to correct the variance of second_id effects corresponding to each observation in obs.

Bii_cov: The Bii to correct the co-variance of first_id and second_id effects corresponding to each observation in obs.

Examples

Suppose we have a dataset my_data.csv that is stored in "C:\\Users\\owner\\Desktop\\vchdfe\\". The structure of the data is such that the fifth column corresponds to the worker identifier and the third column corresponds to the firm identifier (what we refer to first and second identifiers, respectively). Moreover, we want to store a summary of the results here "C:\\Users\\owner\\Desktop\\vchdfe\\summary.txt", and the detailed output (Pii and Bii matrices, stored fixed effects, etc.) here "C:\\Users\\owner\\Desktop\\vchdfe\\output.csv". Important: if the dataset has less than 5000 observations the app will default to use the exact algorithm; otherwise it will compute JLA. See Appendix in KSS for more information.

To obtain all three bias-corrected components (variance of worker effects, variance of firm effects, and covariance of worker-firm effects), we only need to type in the powershell

vchdfe "C:\\Users\\owner\\Desktop\\vchdfe\\my_data.csv" --first_id 5 --second_id 3 --write_results --write_detailed_CSV --detailed_output_path "C:\\Users\\owner\\Desktop\\vchdfe\\output.csv" --results_path "C:\\Users\\owner\\Desktop\\vchdfe\\summary.txt"

To run the same thing but with only 1000 simulations, we type in the powershell

vchdfe "C:\\Users\\owner\\Desktop\\vchdfe\\my_data.csv" --first_id 5 --second_id 3 --write_results  --write_detailed_CSV --detailed_output_path "C:\\Users\\owner\\Desktop\\vchdfe\\output.csv" --results_path "C:\\Users\\owner\\Desktop\\vchdfe\\summary.txt" --simulations 1000

To only obtain the bias-correction for the variance of firm effects, we type in the powershell

Modify this Alim

Functions in this package

Main Functions

VarianceComponentsHDFE.leave_out_estimation — Method

leave_out_estimation(y, first_id, second_id, controls, settings)

Returns the bias-corrected components, the vector of coefficients, the corresponding fixed effects for every observation, and the diagonal matrices containing the Pii and Biis.

Arguments

y: outcome vector

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

settings: settings based on VCHDFESettings

controls: at this version only controls=nothing is supported.

VarianceComponentsHDFE.get_leave_one_out_set — Method

get_leave_one_out_set(y, first_id, second_id, settings, controls)

Returns a tuple with the observation number of the original dataset that belongs to the Leave-out connected set as described in Kline,Saggio, Solvesten. It also provides the corresponding outcome and identifiers in this connected set.

Arguments

y: outcome vector

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

settings: settings based on VCHDFESettings

controls: at this version only controls=nothing is supported.

Auxiliary Functions

VarianceComponentsHDFE.find_connected_set — Method

find_connected_set(y, first_idvar, second_idvar, settings)

Returns a tuple of observation belonging to the largest connected set with the corresponding identifiers and outcomes. This requires to have the data sorted by first identifier, and time period (e.g. we sort by worked id and year). This is also the set where we can run AKM models with the original data.

Arguments

y: outcome (e.g. log wage)

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

settings: settings based on data type VCHDFESettings. Please see the reference provided below.

VarianceComponentsHDFE.prunning_connected_set — Method

prunning_connected_set(yvec, first_idvar, second_idvar, obs_id, settings)

This function prunes the dataset from articulation points. If the first identifier is worker id it means that it prunes workers that would disconnect the graph if they were dropped.

Arguments

yvec: outcome (e.g. log wage)

first_idvar: first identifier (e.g. worker id)

second_idvar: second identifier (e.g. firm id)

obs_id: observation identifier.

settings: settings based on data type VCHDFESettings. Please see the reference provided below.

VarianceComponentsHDFE.drop_single_obs — Method

drop_single_obs(yvec, first_idvar, second_idvar, obs_id)

This function drops observations that correspond with first identifiers with a single observation. For example, if first identifier is worker id, it will drop observations for workers that only appear once in the data.

Arguments

yvec: outcome (e.g. log wage)

first_idvar: first identifier (e.g. worker id)

second_idvar: second identifier (e.g. firm id)

obs_id: observation identifier.

VarianceComponentsHDFE.compute_movers — Method

compute_movers(first_id, second_id)

Returns a vector that indicates whether the first_id (e.g. worker) is a mover across second_id (e.g. firms), as well as a vector with the number of periods that each first_id appears.

Arguments

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

VarianceComponentsHDFE.eff_res — Method

eff_res(_, X, first_id, second_id, match_id, K, settings)

This function computes the diagonal matrices containing Pii and Bii under the Exact Algorithm. See appendix in KSS for more information.

Arguments

X: the design matrix in the linear model.

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

match_id: match identifier between first and second identifier.For example, match can be the identifier of every worker-firm combination.

K: number of covariates in addition to the fixed effects. Currently only 0 is supported.

settings: settings based on data type VCHDFESettings. Please see the reference provided below.

VarianceComponentsHDFE.eff_res — Method

eff_res(lev, X, first_id, second_id, match_id, K, settings)

This function computes the diagonal matrices containing Pii and Bii under Johnson-Linderstrauss Algorithm. See appendix in KSS for more information.

Arguments

X: the design matrix in the linear model.

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

match_id: match identifier between first and second identifier.For example, match can be the identifier of every worker-firm combination.

K: number of covariates in addition to the fixed effects. Currently only 0 is supported.

settings: settings based on data type VCHDFESettings. Please see the reference provided below.

VarianceComponentsHDFE.compute_matchid — Method

compute_matchid(second_id, first_id)

Computes a match identifier for every combination of first and second identifier. For example, this can be the match identifier of worker-firm combinations.

Arguments

first_id: first identifier (e.g. worker id)

second_id: second identifier (e.g. firm id)

Datatypes in this package

VarianceComponentsHDFE.ExactAlgorithm — Type

struct ExactAlgorithm <: AbstractLeverageAlgorithm

Data type to pass to VCHDFESettings type, to indicate Exact algorithm

VarianceComponentsHDFE.JLAAlgorithm — Type

struct JLAAlgorithm <: AbstractLeverageAlgorithm

Data type to pass to VCHDFESettings type, to indicate JLA algorithm

Fields

num_simulations: number of simulations in estimation. If num_simulations = 0, defaults to 100 * log(#total fixed effect)"

VarianceComponentsHDFE.VCHDFESettings — Type

struct VCHDFESettings{LeverageAlgorithm}

The VCHDFESettings type is to pass information to methods regarding which algorithm to use.

Fields

cg_maxiter: maximum number of iterations (default = 300)

leverage_algorithm: which type of algorithm to use (default = ExactAlgorithm())

first_id_effects: includes first id effects. At this version it is required to include the firstideffects. (default = true)

cov_effects: includes covariance of first-second id effects. At this version it is required to include the cov_effects. (default = true)

print_level: prints the state of the program in std output. If print_level = 0, the app prints nothing in the std output. (default = 1)

first_id_display_small: name of the first id in lower cases (default = person)

first_id_display: name of the first id (default = Person)

second_id_display_small: name of the second id in lower cases (default = firm)

second_id_display: name of the second id (default = Firm)

observation_id_display_small: name of the observation id in lower cases (default = wage)

observation_id_display: name of the observation id (default = Wage)

About the current version

The bias-correction currently only runs on a model without controls.

If the user wants, they can manually preadjust on her own the outcome. For instance, in an AKM context, the user can run first $y = pe + fe + Xb + e$ , where $pe$ are person effects and $fe$ are firm effects, and feed into the routine $y-Xb$ as the outcome.